requirements.txt
flask
flask_cors
docker
pyyaml
